修改 IBObject 的 x、y、w、h 为 double 类型，Point 的 x, y 为 long 类型
1.修改 IBObject 的 x、y、w、h 为 double 类型
==>通过修改 Bounds 接口相应方法
    ==>新增 getIntX 等方法兼容 int 值获取
    ==>BaseBouds 相应调整
==>IBCharacter 构造方法 x、y参数对应修改为 double 类型
==>MovableIBCharacter 修改 speed 相关参数为 double 类型

2.Point 的 x, y 为 long 类型
==>四舍五入方式改为使用 Math.round，去掉不再需要的取整强制转换

+++ ib-api/src/main/java/cn/milai/ib/container/ui/Camera.java
@@ -24,7 +24,7 @@ public interface Camera {
 	 * @param x
 	 * @return
 	 */
-	int toRealX(UIContainer c, int x);
+	double toRealX(UIContainer c, double x);
 
 	/**
 	 * 获取指定显示 y 在容器 c 中对应的实际 y 坐标
@@ -32,5 +32,5 @@ public interface Camera {
 	 * @param y
 	 * @return
 	 */
-	int toRealY(UIContainer c, int y);
+	double toRealY(UIContainer c, double y);
 }
